backend/requirements.txt
# Core FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Data Validation and Settings
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0

# Mail Services
aiosmtplib==3.0.1
aiofiles==23.2.1

# SaaS and Multi-tenancy Support
redis==5.0.1
celery==5.3.4
kombu==5.3.4

# Rate Limiting and Caching
slowapi==0.1.9
python-limits==3.6.0

# Monitoring and Logging
structlog==23.2.0
prometheus-client==0.19.0

# Background Tasks and Scheduling
apscheduler==3.10.4
rq==1.15.1

# File Storage and Processing
boto3==1.34.0
pillow==10.1.0

# API Documentation and Validation
pydantic-extra-types==2.2.0
typing-extensions==4.8.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
requests==2.31.0
factory-boy==3.3.0

# Utilities
click==8.1.7
rich==13.7.0
tenacity==8.2.3